% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{compare_models}
\alias{compare_models}
\title{Compare empirical models}
\usage{
compare_models(...)
}
\arguments{
\item{...}{models with argument names indicating the names of models}
}
\description{
Compare empirical models
}
