% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit}
\alias{fit}
\title{Fit a model of the BassSIR family to epidemic data}
\usage{
fit(d, r_rec, r_death, type = c("BassSIR", "SIR", "Growth"), hyper,
  n_iter = 10000, ...)
}
\arguments{
\item{d}{a BassSIR data}

\item{r_rec}{recovery rate}

\item{r_death}{death rate}

\item{type}{type of model}

\item{n_iter}{number of interation to keep}

\item{...}{arguments for jags(...)}
}
\description{
Fit a model of the BassSIR family to epidemic data
}
