% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_full}
\alias{fit_full}
\title{Fit a full model of the BassSIR family to epidemic data}
\usage{
fit_full(d, r_iso = 1/6.1, type = c("BassSIR", "SIR", "Growth"), hyper,
  n_iter = 10000, n_collect = 1000, ...)
}
\arguments{
\item{d}{a BassSIR data}

\item{r_iso}{rate to isolation}

\item{type}{type of model}

\item{n_iter}{number of interation}

\item{n_collect}{number of interation to collect}

\item{...}{arguments for jags(...)}
}
\description{
Fit a full model of the BassSIR family to epidemic data
}
