% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.EstBassSIR}
\alias{fitted.EstBassSIR}
\alias{summary.fittedBassSIR}
\title{Fitted values of BassSIR models}
\usage{
\method{fitted}{EstBassSIR}(est, method = c("match", "forward"))

\method{summary}{fittedBassSIR}(f)
}
\arguments{
\item{method}{}
}
\description{
Fitted values of BassSIR models
}
