% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_covid19}
\alias{n_covid19}
\title{Confirmed cases, the recovered, and peopole who died from COVID-19}
\format{A list of data frames Province -> data.frame:
\describe{
  \item{Date}{Date of the entry}
  \item{Confirmed}{Confirmed cases}
  \item{Dead}{People who died from COVID-19}
  \item{Cured}{People who recovered from COVID-19}
  ...
}}
\source{
\url{}
}
\usage{
n_covid19
}
\description{
A dataset containing the Confirmed cases, the recovered, and peopole who died from COVID-19
by chinese province
}
\keyword{datasets}
