% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{run_scenario}
\alias{run_scenario}
\alias{compare_scenarios}
\title{Run scenario analysis}
\usage{
run_scenario(sim, fn)

compare_scenarios(sim, ..., fn_change = c("Y0", "Y1", "Yt"))
}
\arguments{
\item{sim}{baseline simulations}

\item{fn}{a function for modifying parameters}
}
\description{
Run scenario analysis
}
\examples{
library(R2jags)
cases <- as_bass_data(n_covid19$Hubei)
est <- BassSIR::fit(cases, r_rec = 1/22.2, r_death = 1/22.3, type = "BassSIR")

sim <- simulate(est, nsim = 500)

zero_kappa <- function(pars) {
  pars$kappa <- rep(0, length(pars$kappa))
  return(pars)
}

lockdown <- run_scenario(sim, zero_kappa)

cp <- compare_scenarios(sim, Lockdown = lockdown)
}
