% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stats_fn}
\alias{stats_fn}
\title{Calculate bootstrap credible interval}
\usage{
stats_fn(x, digits = getOption("qwraps2_frmt_digits", 2))
}
\arguments{
\item{digits}{}
}
\description{
Calculate bootstrap credible interval
}
