% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{summary.simBassSIR}
\alias{summary.simBassSIR}
\alias{print.summarySimBassSIR}
\title{Summarise epidemiological variables from simulation results}
\usage{
\method{summary}{simBassSIR}(sim, i = 1)

\method{print}{summarySimBassSIR}(obj)
}
\arguments{
\item{sim}{a simulation result}

\item{i}{an indicator specifying time points, 1 as the default}
}
\description{
Summarise epidemiological variables from simulation results
}
