% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{to_vis_trajectory}
\alias{to_vis_trajectory}
\title{Combine information for visualise prevalence}
\usage{
to_vis_trajectory(est, sim, traj = c("I", "Rt"), method = "match")
}
\arguments{
\item{est}{model estimates}

\item{sim}{model simulations}

\item{method}{method for finding fitted values}
}
\description{
Combine information for visualise prevalence
}
