% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{as_bass_data}
\alias{as_bass_data}
\title{Prepare data for BassSIR modelling}
\usage{
as_bass_data(d, location = "Place X", i_date = "Date",
  i_confirmed = "Confirmed", i_recovered = "Cured", i_dead = "Dead")
}
\arguments{
\item{d}{source data.frame}

\item{location}{identity of the data}

\item{i_date}{index indicating date}

\item{i_confirmed}{index indicating confirmed cases}

\item{i_recovered}{index indicating recovered people}

\item{i_dead}{index indicating the number of dead people}
}
\description{
Prepare data for BassSIR modelling
}
\examples{
as_bass_data(n_covid19$Hubei)

}
