% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit}
\alias{fit}
\title{Fit a BassSIR family model to epidemic data}
\usage{
fit(d, r_rec, r_death, type = c("BassSIR", "SIR"), n_iter = 10000, ...)
}
\arguments{
\item{...}{}
}
\description{
Fit a BassSIR family model to epidemic data
}
