% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.estBassSIR}
\alias{fitted.estBassSIR}
\title{Fitted values of BassSIR models}
\usage{
\method{fitted}{estBassSIR}(est, method = c("match", "forward"))
}
\arguments{
\item{method}{}
}
\description{
Fitted values of BassSIR models
}
