% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{summary.estBassSIR}
\alias{summary.estBassSIR}
\alias{print.summaryEstBassSIR}
\title{Summarise a fitted Bass SIR model}
\usage{
\method{summary}{estBassSIR}(est)

\method{print}{summaryEstBassSIR}(est)
}
\arguments{
\item{est}{}
}
\description{
Summarise a fitted Bass SIR model
}
